Fixed config parsing to take defaults into account. Also changed module cache to be uint 16 (not sure why it was uint8)

+++ cosmwasm/enclaves/ffi-types/src/types.rs
@@ -40,7 +40,7 @@ pub struct RuntimeConfiguration {
     /// The amount of wasmi modules cached in an LRU cache inside the enclave.
     /// This speeds up the execution of recently used modules, but has a significant
     /// memory overhead.
-    pub module_cache_size: u8,
+    pub module_cache_size: u32,
 }
 
 /// This struct holds a pointer to memory in userspace, that contains the storage
